requirements.txt
Django>=5.1,<5.2
python-dotenv>=1.0.0
social-auth-app-django>=4.0.0
django-sslserver>=0.22
requests>=2.28
gunicorn>=20.0.4
psycopg2-binary>=2.9
PyPDF2>=2.10.0
python-docx>=0.8.11
langchain>=0.0.200
langchain-google-genai>=0.0.5
faiss-cpu>=1.7.1